% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_heatmap.R
\name{hierarchical_colors}
\alias{hierarchical_colors}
\title{Color tissues/cell types hierarchically}
\usage{
hierarchical_colors(mat_meta)
}
\arguments{
\item{mat_meta}{A \code{data.frame} with columns Tissue, CellDeriv,
Cell, and TF.}
}
\value{
A named character vector of colors covering all unique values
  of Tissue, CellDeriv, Cell, and TF.
}
\description{
Use the same base color for groups of tissues/cell types
but vary the brightness within subgroups. Requires \pkg{pals}
and \pkg{shades}.
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
\keyword{internal}
