% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_enrichment.R
\name{IMPACT_compute_enrichment}
\alias{IMPACT_compute_enrichment}
\title{Compute enrichment of IMPACT scores}
\usage{
IMPACT_compute_enrichment(annot_melt, locus = NULL)
}
\arguments{
\item{annot_melt}{A melted \code{data.table} of IMPACT annotations
with columns including IMPACT_score, SNP, TF, Tissue, Cell, CellDeriv,
and various fine-mapping result columns.}

\item{locus}{Optional locus name to add to the output.}
}
\value{
A \code{data.table} of enrichment results per SNP group and
  annotation.
}
\description{
Conduct IMPACT enrichment between SNP groups and fine-mapping methods.
Enrichment is computed as the ratio of IMPACT signal in a SNP group
to the proportion of SNPs in that group.
}
\examples{
\dontrun{
enrich <- IMPACT_compute_enrichment(annot_melt = annot_melt,
                                    locus = "BST1")
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
