% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMPACT_files}
\alias{IMPACT_files}
\title{IMPACT files}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 88 rows and 9 columns.
}
\source{
\code{
IMPACT_files <- downloadR::zenodo_list(token = token)[!endsWith(filename,
                                                     c(".zip",".tbi")),]
usethis::use_data(IMPACT_files, overwrite = TRUE)
}
}
\usage{
data("IMPACT_files")
}
\description{
URLs to annotations/LD-scores generated by
\href{https://github.com/immunogenomics/IMPACT}{IMPACT}
(Inference and Modeling of Phenotype-related ACtive Transcription).
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
\keyword{datasets}
