% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_get_annotation_key.R
\name{IMPACT_get_annotation_key}
\alias{IMPACT_get_annotation_key}
\title{Get \emph{IMPACT} annotation key}
\usage{
IMPACT_get_annotation_key(
  URL = paste0("https://github.com/immunogenomics/IMPACT/",
    "raw/master/IMPACT707/IMPACT_annotation_key.txt"),
  save_path = "./IMPACT/IMPACT_annotation_key.txt.gz",
  save_key = FALSE,
  force_new_download = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{URL}{URL to the IMPACT annotation key file.}

\item{save_path}{Local path to save the annotation key.}

\item{save_key}{Whether to save the annotation key locally.}

\item{force_new_download}{Whether to force re-downloading even if
a local copy exists.}

\item{verbose}{Print messages.}
}
\value{
A \code{data.table} with annotation metadata.
}
\description{
Includes the study source, tissue, cell type/line,
and cell subtype of each of the 500 annotations in \emph{IMPACT}.
}
\examples{
\dontrun{
annot_key <- IMPACT_get_annotation_key(save_key = FALSE)
head(annot_key)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
