% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_get_annotations.R
\name{IMPACT_get_annotations}
\alias{IMPACT_get_annotations}
\title{Download \emph{IMPACT} annotations}
\usage{
IMPACT_get_annotations(
  baseURL = paste0("https://github.com/immunogenomics/IMPACT/",
    "raw/master/IMPACT707/Annotations"),
  chrom = NULL,
  dat = NULL,
  nThread = 1,
  all_snps_in_range = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{baseURL}{Base URL for IMPACT annotation files.}

\item{chrom}{Chromosome number.}

\item{dat}{A \code{data.table} with at least columns CHR, SNP, POS.}

\item{nThread}{Number of threads for \code{data.table::fread}.}

\item{all_snps_in_range}{Whether to include all SNPs in the range
(not just those in \code{dat}).}

\item{verbose}{Print messages.}
}
\value{
A melted \code{data.table} of IMPACT annotations
  merged with annotation metadata.
}
\description{
Includes the raw annotation itself,
as well as per-SNP \emph{IMPACT} scores for each annotation.
}
\details{
Unfortunately, you have to download the entire chromosome file at once,
 because they aren't Tabix indexed. To minimize the memory load,
 this function only keeps the portion of the \emph{IMPACT} file that
 overlaps with the coordinates in \code{dat}.
}
\examples{
\dontrun{
BST1 <- echodata::BST1
annot_melt <- IMPACT_get_annotations(dat = BST1)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
