% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_get_ldscores.R
\name{IMPACT_get_ldscores}
\alias{IMPACT_get_ldscores}
\title{Download LD scores from IMPACT publication}
\source{
T Amariuta et al., IMPACT: Genomic Annotation of Cell-State-Specific
Regulatory Elements Inferred from the Epigenome of Bound Transcription
Factors. The American Journal of Human Genetics, 1-17 (2019).
}
\usage{
IMPACT_get_ldscores(chrom = NULL, dat = NULL, nThread = 1)
}
\arguments{
\item{chrom}{Chromosome number.}

\item{dat}{A \code{data.table} with at least columns CHR, SNP, POS.
If provided, \code{chrom} is extracted from the first row.}

\item{nThread}{Number of threads for \code{data.table::fread}.}
}
\value{
A \code{data.table} of LD scores, optionally merged with
  \code{dat}.
}
\description{
Downloads per-SNP LD scores from LD Score regression,
first conducted in the original IMPACT publication.
}
\examples{
\dontrun{
BST1 <- echodata::BST1
ldscore <- IMPACT_get_ldscores(dat = BST1)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
