% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_postprocess.R
\name{IMPACT_get_top_annotations}
\alias{IMPACT_get_top_annotations}
\title{Get the top annotation(s)}
\usage{
IMPACT_get_top_annotations(
  ANNOT_MELT,
  snp.filter = "!is.na(IMPACT_score)",
  top_annotations = 1,
  force_one_annot_per_locus = FALSE
)
}
\arguments{
\item{ANNOT_MELT}{A melted \code{data.table} of IMPACT annotations.}

\item{snp.filter}{An expression (as character) used to filter SNPs
when computing summary statistics.}

\item{top_annotations}{Number of top annotations to keep per locus,
or \code{FALSE} to keep all.}

\item{force_one_annot_per_locus}{Whether to force exactly one annotation
per locus.}
}
\value{
A \code{data.table} with top annotations per locus.
}
\description{
Get the annotation(s) with the top mean \emph{IMPACT} for a given set
of SNPs.
}
\examples{
\dontrun{
top_impact <- IMPACT_get_top_annotations(ANNOT_MELT = ANNOT_MELT)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
