% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_enrichment.R
\name{IMPACT_iterate_enrichment}
\alias{IMPACT_iterate_enrichment}
\title{Iterate IMPACT enrichment tests}
\usage{
IMPACT_iterate_enrichment(ANNOT_MELT, verbose = TRUE)
}
\arguments{
\item{ANNOT_MELT}{A melted \code{data.table} of IMPACT annotations
that must include a \code{Locus} column.}

\item{verbose}{Print messages.}
}
\value{
A \code{data.table} of enrichment results across all loci.
}
\description{
Run \code{\link{IMPACT_compute_enrichment}} across all unique loci
in the ANNOT_MELT dataset.
}
\examples{
\dontrun{
ENRICH <- IMPACT_iterate_enrichment(ANNOT_MELT = ANNOT_MELT)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
