% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_get_annotations.R
\name{IMPACT_iterate_get_annotations}
\alias{IMPACT_iterate_get_annotations}
\title{Gather all \emph{IMPACT} annotations that overlap with your query}
\usage{
IMPACT_iterate_get_annotations(
  merged_DT,
  IMPACT_score_thresh = 0.1,
  baseURL = paste0("https://github.com/immunogenomics/IMPACT/",
    "raw/master/IMPACT707/Annotations"),
  all_snps_in_range = TRUE,
  top_annotations = FALSE,
  force_one_annot_per_locus = FALSE,
  snp.filter = "!is.na(SNP)",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{merged_DT}{A merged \code{data.table} containing fine-mapping results
across multiple loci with at least columns CHR and Locus.}

\item{IMPACT_score_thresh}{Minimum IMPACT score threshold for filtering.}

\item{baseURL}{Base URL for IMPACT annotation files.}

\item{all_snps_in_range}{Whether to include all SNPs in range.}

\item{top_annotations}{Number of top annotations to keep,
or \code{FALSE} to keep all.}

\item{force_one_annot_per_locus}{Whether to force one annotation per locus.}

\item{snp.filter}{An expression (as character) used to filter SNPs
for top annotation selection.}

\item{nThread}{Number of threads for data.table operations.}

\item{verbose}{Print messages.}
}
\value{
A \code{data.table} of IMPACT annotations across all chromosomes.
}
\description{
Iterates over each unique locus.
\bold{\emph{WARNING!}} These files are quite large and you need to make sure
you have ample memory and storage on your computer for best results.
}
\examples{
\dontrun{
data("merged_DT")
ANNOT_MELT <- IMPACT_iterate_get_annotations(merged_DT = merged_DT)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
