% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_plot.R
\name{IMPACT_plot_enrichment}
\alias{IMPACT_plot_enrichment}
\title{Plot IMPACT score enrichment}
\usage{
IMPACT_plot_enrichment(ENRICH, show_plot = TRUE)
}
\arguments{
\item{ENRICH}{A \code{data.table} of enrichment results
as returned by \code{\link{IMPACT_compute_enrichment}} or
\code{\link{IMPACT_iterate_enrichment}}.}

\item{show_plot}{Whether to display the plots.}
}
\value{
A list of \code{ggplot} objects (invisibly).
}
\description{
Visualize IMPACT enrichment results with bar plots and violin plots.
}
\examples{
\dontrun{
plots <- IMPACT_plot_enrichment(ENRICH = ENRICH)
}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
