% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT_heatmap.R
\name{prepare_mat_meta}
\alias{prepare_mat_meta}
\title{Prepare data matrix and metadata for \pkg{ComplexHeatmap}}
\usage{
prepare_mat_meta(
  TOP_IMPACT,
  snp.group = "Consensus",
  value.var = "mean_IMPACT",
  fill_na = 0
)
}
\arguments{
\item{TOP_IMPACT}{A \code{data.table} of top IMPACT annotations.}

\item{snp.group}{Name of the SNP group to use for metadata merge.}

\item{value.var}{Name of the value column to use in the cast.}

\item{fill_na}{Value to use for filling NA entries.}
}
\value{
A \code{data.frame} with loci as rows and SNP groups as columns,
  plus metadata columns.
}
\description{
Prepare data matrix and metadata for \pkg{ComplexHeatmap}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT_compute_enrichment}()},
\code{\link{IMPACT_files}},
\code{\link{IMPACT_get_annotation_key}()},
\code{\link{IMPACT_get_annotations}()},
\code{\link{IMPACT_get_ldscores}()},
\code{\link{IMPACT_get_top_annotations}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{IMPACT_iterate_enrichment}()},
\code{\link{IMPACT_iterate_get_annotations}()},
\code{\link{IMPACT_plot_enrichment}()},
\code{\link{IMPACT_plot_impact_score}()},
\code{\link{IMPACT_postprocess_annotations}()},
\code{\link{IMPACT_process}()},
\code{\link{IMPACT_query}()},
\code{\link{IMPACT_snp_group_boxplot}()},
\code{\link{hierarchical_colors}()}
}
\concept{IMPACT}
\keyword{internal}
