% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLICEAI_query.R
\name{SPLICEAI_query_vcf}
\alias{SPLICEAI_query_vcf}
\title{Query genome-wide \emph{SpliceAI} results (VCF format)}
\source{
\url{https://github.com/Illumina/SpliceAI}
\doi{10.1016/j.cell.2018.12.015}
}
\usage{
SPLICEAI_query_vcf(dat, precomputed_path, subset_vcf = "subset.vcf")
}
\arguments{
\item{dat}{A \code{data.table} with at least columns \code{CHR}
(chromosome) and \code{POS} (base-pair position).}

\item{precomputed_path}{Path to the bgzipped and tabix-indexed
pre-computed SpliceAI VCF file.}

\item{subset_vcf}{Path to write the subset VCF output.}
}
\value{
The exit code from the \code{bcftools} system call
  (invisibly).
}
\description{
Subset a genome-wide pre-computed \emph{SpliceAI} VCF file using
\code{bcftools} for the region covered by the supplied data.
}
\examples{
\dontrun{
dat <- data.table::data.table(CHR = 1, POS = 100000:100500)
SPLICEAI_query_vcf(
    dat = dat,
    precomputed_path = "spliceai_scores.vcf.gz",
    subset_vcf = "subset.vcf"
)
}
}
\seealso{
Other SPLICEAI: 
\code{\link{SPLICEAI_plot}()},
\code{\link{SPLICEAI_query_api}()},
\code{\link{SPLICEAI_query_tsv}()},
\code{\link{SPLICEAI_query_tsv_iterate}()},
\code{\link{SPLICEAI_run}()},
\code{\link{SPLICEAI_snp_probs}()}
}
\concept{SPLICEAI}
